split telegram into receiving and sending
The telegram receiving can be run in a separate thread to not block
the main thread, while sending must be easily possible from other
places in the system.
